Add PrimitiveCodec interface
