+++ tablet/src/main/scala/wvlet/airframe/tablet/msgpack/MessageCodec.scala
@@ -1,10 +1,11 @@
 package wvlet.airframe.tablet.msgpack
 
 import org.msgpack.core.{MessagePack, MessagePacker, MessageUnpacker}
-import wvlet.airframe.tablet.msgpack.MessageCodec.{ErrorCode, INVALID_DATA}
+import org.msgpack.value.Value
+import wvlet.airframe.tablet.msgpack.MessageCodec.ErrorCode
 
 import scala.reflect.runtime.{universe => ru}
-import scala.util.{Failure, Try}
+import scala.util.{Failure, Success, Try}
 
 trait MessageCodec[A] {
   def pack(p: MessagePacker, v: A)
@@ -13,14 +14,14 @@ trait MessageCodec[A] {
   // TODO add specialized methods for primitive values
   // def unpackInt(u:MessageUnpacker) : Int
 
-  def pack(v: A): Array[Byte] = {
+  def packToBytes(v: A): Array[Byte] = {
     val packer = MessagePack.newDefaultBufferPacker()
     pack(packer, v)
     packer.toByteArray
   }
 
-  def unpack(data: Array[Byte]): Option[A] = unpack(data, 0, data.length)
-  def unpack(data: Array[Byte], offset: Int, len: Int): Option[A] = {
+  def unpackBytes(data: Array[Byte]): Option[A] = unpackBytes(data, 0, data.length)
+  def unpackBytes(data: Array[Byte], offset: Int, len: Int): Option[A] = {
     val unpacker = MessagePack.newDefaultUnpacker(data, offset, len)
     val v        = new MessageHolder
     unpack(unpacker, v)
@@ -28,6 +29,27 @@ trait MessageCodec[A] {
       None
     } else {
       Some(v.getLastValue.asInstanceOf[A])
+    }
+  }
+}
+
+trait MessageValueCodec[A] extends MessageCodec[A] {
+  def pack(v: A): Value
+  def unpack(v: Value): A
+
+  override def pack(p: MessagePacker, v: A) {
+    p.packValue(pack(v))
+  }
+
+  override def unpack(u: MessageUnpacker, v: MessageHolder) {
+    val vl = u.unpackValue()
+    Try(unpack(vl)) match {
+      case Success(x) =>
+        // TODO tell the value type of the object to MessageHolder
+        // We cannot use MessageValueCodec for primitive types, which uses v.getInt, v.getString, etc.
+        v.setObject(x)
+      case Failure(e) =>
+        v.setError(e)
     }
   }
 }
